+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 2.0.17
+- REMReM lookups controlled 'failIfNoneFound' and 'failIfMultipleFound' lookups per lookup
+  object within an event instead of globally per call.
 - Removed Ericsson specific configuration from github.
 
 ## 2.0.16